[FIX] account_edi: search partner based on VAT with spaces
This solution is not perfect. In order to be, we would need to have a
sanitized function that is used to store the number in the database, and
to use the same function to search in it. This will most likely be done
in master with a refactoring of `base_vat`, on which `account` will
depend one way or another.

In the meantime, we need to support cases that were working before these
fixes:
https://github.com/odoo/odoo/commit/75826c168556e9fe085164ac424424542422202a
https://github.com/odoo/odoo/commit/a7f28fa64235f1085432ddccd3090f15be938ef7

Since these, it was indeed impossible to detect a partner based on his
VAT for Swiss partners if `base_vat` was installed, which is the
default.

closes odoo/odoo#81250

X-original-commit: d718ac60cc73738ec0a48944eb61102837cf088c
Signed-off-by: Olivier Colson <oco@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>