[FIX] pos_cash_rounding: Fix value under rounding
It should be allowed to pay amount that are not rounded if the value is less than the rounding value.
Ex:
Rounding value: 0.05
Amount to pay: 0.04
The amount to pay should remain 0.04 and not be rounded to 0.05.

This fix allows the user to pay for values under the rounding value without applying the rounding.

closes odoo/odoo#81365

X-original-commit: 573d0672127cea22ba41dd9c290377d0f72538a6
Signed-off-by: Masereel Pierre <pim@odoo.com>