[FIX] l10n_sa_invoice: relax constraint for delivery date
We had a constraint(raise) that said that the delivery date needed
to be after the invoice date, but this is putting too many
constraints.

However, we put a help message to say that they should put the
date of the last delivery.  Also, the delivery date should not be
changed once confirmed like the other information in the invoice.

X-original-commit: 15badb424dce17bfc3e7cc25a4d708576d4fe6b5
Part-of: odoo/odoo#81611