[FIX] l10n_es_edi_sii: certificate problems with newer libraries
In urllib3 1.25, an _is_key_file_encrypted method was added
where urllib3 checks if the keyfile is encrypted or not, but
urllib3 uses filenames, while we want to use the contents of cert.
So, in our case, this gave a traceback because it could not open
the file.

The final simple solution for now is to keep keyfile empty as
we store our info in certfile anyways already.

X-original-commit: 6588f85ba6d159e06ab7b955d34d783a1ad1da35
Part-of: odoo/odoo#81908